+++ NLE/NLCore/NL_Distributor.h
@@ -13,9 +13,9 @@ namespace NLE
 			class Distributor
 			{
 			public:
-				virtual void distributeFrom(uint_fast8_t sysId) = 0;
-				virtual void distributeTo(uint_fast8_t sysId) = 0;
-				virtual std::vector<uint_fast8_t>& getEndpoints() = 0;
+				virtual void distributeFrom(uint_fast32_t sysId) = 0;
+				virtual void distributeTo(uint_fast32_t sysId) = 0;
+				virtual std::vector<uint_fast32_t>& getEndpoints() = 0;
 				virtual void processRequests(){}
 			};
 		}
