+++ NLE/NLCore/NL_SysManager.h
@@ -26,15 +26,15 @@ namespace NLE
 				std::unique_ptr<IEngine> const& iEngine);
 			void release();
 
-			uint_fast8_t getNumSystems();
-			ExecutionDesc& getExecutionDesc(uint_fast8_t sysId);
+			uint_fast32_t getNumSystems();
+			ExecutionDesc& getExecutionDesc(uint_fast32_t sysId);
 
-			std::unique_ptr<System> const& getSystem(uint_fast8_t sysId) const;
+			std::unique_ptr<System> const& getSystem(uint_fast32_t sysId) const;
 			void attachSystem(ExecutionDesc execDesc, std::unique_ptr<System> system);
 		private:
-			std::unordered_map<uint_fast8_t, std::unique_ptr<System>> _systems;
-			std::unordered_map<uint_fast8_t, ExecutionDesc> _executionDesc;
-			uint_fast8_t _numSystems;
+			std::unordered_map<uint_fast32_t, std::unique_ptr<System>> _systems;
+			std::unordered_map<uint_fast32_t, ExecutionDesc> _executionDesc;
+			uint_fast32_t _numSystems;
 		};
 	}
 }
